+++ src/main/java/uk/gov/di/ipv/cri/kbv/api/domain/Control.java
@@ -1,4 +1,4 @@
-package uk.gov.di.cri.experian.kbv.api.domain;
+package uk.gov.di.ipv.cri.kbv.api.domain;
 
 public class Control {
     private String authRefNo;
