+++ src/main/java/uk/gov/di/ipv/cri/kbv/api/domain/PersonIdentity.java
@@ -1,4 +1,4 @@
-package uk.gov.di.cri.experian.kbv.api.domain;
+package uk.gov.di.ipv.cri.kbv.api.domain;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
