+++ src/main/java/uk/gov/di/ipv/cri/kbv/api/domain/QuestionAnswerResponse.java
@@ -1,7 +1,6 @@
-package uk.gov.di.cri.experian.kbv.api.domain;
+package uk.gov.di.ipv.cri.kbv.api.domain;
 
-public class QuestionsResponse {
-
+public class QuestionAnswerResponse {
     protected Control control;
     protected Questions questions;
     protected Results results;
@@ -11,31 +10,31 @@ public class QuestionsResponse {
         return control;
     }
 
-    public void setControl(Control control) {
-        this.control = control;
+    public void setControl(Control value) {
+        this.control = value;
     }
 
     public Questions getQuestions() {
         return questions;
     }
 
-    public void setQuestions(Questions questions) {
-        this.questions = questions;
+    public void setQuestions(Questions value) {
+        this.questions = value;
     }
 
     public Results getResults() {
         return results;
     }
 
-    public void setResults(Results results) {
-        this.results = results;
+    public void setResults(Results value) {
+        this.results = value;
     }
 
     public Error getError() {
         return error;
     }
 
-    public void setError(Error error) {
-        this.error = error;
+    public void setError(Error value) {
+        this.error = value;
     }
 }
