+++ src/main/java/uk/gov/di/ipv/cri/kbv/api/handler/SessionHandler.java
@@ -1,4 +1,4 @@
-package uk.gov.di.cri.experian.kbv.api.handler;
+package uk.gov.di.ipv.cri.kbv.api.handler;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
@@ -9,9 +9,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import software.amazon.lambda.powertools.tracing.CaptureMode;
 import software.amazon.lambda.powertools.tracing.Tracing;
-import uk.gov.di.cri.experian.kbv.api.domain.PersonIdentity;
-import uk.gov.di.cri.experian.kbv.api.domain.QuestionState;
-import uk.gov.di.cri.experian.kbv.api.service.StorageService;
+import uk.gov.di.ipv.cri.kbv.api.domain.PersonIdentity;
+import uk.gov.di.ipv.cri.kbv.api.domain.QuestionState;
+import uk.gov.di.ipv.cri.kbv.api.service.StorageService;
 
 import java.util.Map;
 import java.util.UUID;
